requirements.txt
PyYAML==6.0.1
jinja2==3.1.3
rich==13.7.0
tabulate==0.9.0
streamlit==1.31.0
pandas==2.1.4
